refactor(objectives): move Source extraction into analysis
Why:
- Confirmed-Objective Source inspection, table repair, grounding, and later paper reconstruction were hidden inside one service method, obscuring the scientific stage boundary.

What:
- Move source-local draft extraction to its direct analysis owner and make the coordinator sequence extraction before scope enrichment, same-paper binding, and pairwise comparison.

Impact:
- Prompt, Source ordering, Evidence identity, progress, fallback, and technical failure behavior remain unchanged, with no compatibility import or forwarding path.

Verification:
- All 120 Objective extraction, routing, comparison, workflow, and transient-model regression tests pass.
