+++ backend/application/core/objectives/analysis/README.md
@@ -11,9 +11,13 @@ Read the analysis responsibilities in real research order:
 2. `evidence_routing.py` turns screened Source units into the transient
    inspection queue consumed by Source-local fact extraction. A route is an
    instruction to inspect a Source, not a scientific finding.
-3. Source extraction, validation, within-paper experiment reconstruction,
-   Evidence materialization, and cross-paper Finding synthesis currently remain
-   in their existing owners until their responsibility slices move here.
+3. `source_extraction.py` inspects each routed Source and produces transient,
+   source-local `ExtractedEvidenceDraft` records. It owns text and table payload
+   construction, structural table repair, deterministic table parsing, model
+   extraction, current Source grounding, and route-scoped technical failures.
+4. Within-paper experiment reconstruction, Evidence materialization, and
+   cross-paper Finding synthesis remain in their existing owners until their
+   responsibility slices move here.
 
 `source_screening.py` owns complete Source-unit accounting, bounded framing
 batches, prompt preflight, model/repaired/fallback dispositions, and frame
@@ -24,6 +28,12 @@ screening decision as scientific Evidence.
 ordering, model or deterministic route decisions, selection hints, and the
 document round-robin extraction queue. It preserves Source identity and cannot
 redirect a route to another paper or Source.
+
+`source_extraction.py` owns the inspection of one exact Source at a time. A
+schema-valid model response is still only a draft: unsupported results abstain,
+unsupported comparison fields are retained only as descriptive evidence, and
+provider or irrecoverable structured-output failures remain technical failed
+drafts. Same-paper Methods/Results binding occurs only after this stage.
 
 Technical JSON parsing, provider retries, prompt schemas, and token accounting
 support this process but do not define its scientific order.
